+++ app/domain/authentication/authn_azure/azure_resource.rb
@@ -1,0 +1,17 @@
+module Authentication
+  module AuthnAzure
+
+    class AzureResource
+      attr_reader :type, :value
+
+      def initialize(type:, value:)
+        @type = type
+        @value = value
+      end
+
+      def ==(other)
+        @type == other.type && @value == other.value
+      end
+    end
+  end
+end
