+++ app/domain/authentication/authn_k8s/k8s_resource.rb
@@ -1,0 +1,17 @@
+module Authentication
+  module AuthnK8s
+
+    class K8sResource
+      attr_reader :type, :value
+
+      def initialize(type:, value:)
+        @type = type
+        @value = value
+      end
+
+      def ==(other)
+        @type == other.type && @value == other.value
+      end
+    end
+  end
+end
