+++ cucumber/authenticators_azure/features/authn_azure_hosts.feature
@@ -54,7 +54,7 @@ Feature: Azure Authenticator - Different Hosts can authenticate with Azure authe
     Then it is unauthorized
     And The following appears in the log after my savepoint:
     """
-    Errors::Authentication::AuthnAzure::RoleMissingConstraint: CONJ00057E Role does not have the required constraint: resource-group
+    Errors::Authentication::RoleMissingConstraint: CONJ00057E Role does not have the required constraint: resource-group
     """
 
   Scenario: Host without subscription-id annotation is denied
@@ -67,7 +67,7 @@ Feature: Azure Authenticator - Different Hosts can authenticate with Azure authe
     Then it is unauthorized
     And The following appears in the log after my savepoint:
     """
-    Errors::Authentication::AuthnAzure::RoleMissingConstraint: CONJ00057E Role does not have the required constraint: subscription-id
+    Errors::Authentication::RoleMissingConstraint: CONJ00057E Role does not have the required constraint: subscription-id
     """
 
   Scenario: Host without any Azure annotation is denied
@@ -79,7 +79,7 @@ Feature: Azure Authenticator - Different Hosts can authenticate with Azure authe
     Then it is unauthorized
     And The following appears in the log after my savepoint:
     """
-    Errors::Authentication::AuthnAzure::RoleMissingConstraint
+    Errors::Authentication::RoleMissingConstraint
     """
 
   Scenario: Host with both identity Azure annotations is denied
@@ -109,7 +109,7 @@ Feature: Azure Authenticator - Different Hosts can authenticate with Azure authe
     Then it is unauthorized
     And The following appears in the log after my savepoint:
     """
-    Errors::Authentication::AuthnAzure::InvalidApplicationIdentity
+    Errors::Authentication::AuthnAzure::InvalidResourceRestrictions
     """
 
   Scenario: Host with incorrect resource-group Azure annotation is denied
@@ -123,7 +123,7 @@ Feature: Azure Authenticator - Different Hosts can authenticate with Azure authe
     Then it is unauthorized
     And The following appears in the log after my savepoint:
     """
-    Errors::Authentication::AuthnAzure::InvalidApplicationIdentity
+    Errors::Authentication::AuthnAzure::InvalidResourceRestrictions
     """
 
   Scenario: Host with incorrect user-assigned-identity annotation is denied
@@ -138,7 +138,7 @@ Feature: Azure Authenticator - Different Hosts can authenticate with Azure authe
     Then it is unauthorized
     And The following appears in the log after my savepoint:
     """
-    Errors::Authentication::AuthnAzure::InvalidApplicationIdentity
+    Errors::Authentication::AuthnAzure::InvalidResourceRestrictions
     """
 
   Scenario: Host with incorrect system-assigned-identity annotation is denied
@@ -153,7 +153,7 @@ Feature: Azure Authenticator - Different Hosts can authenticate with Azure authe
     Then it is unauthorized
     And The following appears in the log after my savepoint:
     """
-    Errors::Authentication::AuthnAzure::InvalidApplicationIdentity
+    Errors::Authentication::AuthnAzure::InvalidResourceRestrictions
     """
 
   Scenario: Non-existing host is denied
