+++ cucumber/kubernetes/features/login.feature
@@ -1,4 +1,4 @@
-Feature: A permitted Conjur host can login with a valid application identity
+Feature: A permitted Conjur host can login with a valid resource restrictions
   that is defined in the id
 
   Scenario: Login as the namespace a pod belongs to.
